+++ crates/ruma-macros/src/events/event_enum.rs
@@ -176,6 +176,11 @@ impl EventEnumKind {
         format_ident!("{variation}{self}Content")
     }
 
+    /// Get the name of the `Custom{kind}Content` type for this kind.
+    fn to_custom_content_ident(self) -> syn::Ident {
+        format_ident!("Custom{self}Content")
+    }
+
     /// Get the list of variations for an event type (struct or enum) for this kind.
     fn event_variations(self) -> &'static [EventVariation] {
         if let Some(common_kind) = self.common_kind() {
@@ -310,11 +315,7 @@ impl EventEnumEntry {
     }
 
     /// Get or generate the path of the event content type for this entry.
-    fn to_event_content_path(
-        &self,
-        kind: EventEnumKind,
-        variation: EventContentVariation,
-    ) -> syn::Path {
+    fn to_event_content_path(&self, kind: EventEnumKind) -> syn::Path {
         let type_prefix = match kind {
             EventEnumKind::ToDevice => "ToDevice",
             // Special case encrypted state event for MSC4362.
@@ -329,7 +330,7 @@ impl EventEnumEntry {
             _ => "",
         };
 
-        let content_name = format_ident!("{variation}{type_prefix}{}EventContent", self.ident);
+        let content_name = format_ident!("{type_prefix}{}EventContent", self.ident);
 
         let mut path = self.ev_path.clone();
         path.segments.push(content_name.into());
@@ -381,47 +382,5 @@ impl EventEnumEntry {
         }
 
         doc
-    }
-}
-
-/// The supported variations for `Any*EventContent` enums.
-#[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-enum EventContentVariation {
-    /// The non-redacted version of the event content, `Any{kind}EventContent`.
-    Original,
-    /// The possibly redacted version of the event content, `AnyPossiblyRedacted{kind}EventContent`.
-    PossiblyRedacted,
-}
-
-impl EventContentVariation {
-    /// Get the event content enum variation matching the given event variation.
-    fn from_event_variation(event_variation: EventVariation) -> Option<Self> {
-        match event_variation {
-            EventVariation::None | EventVariation::Sync | EventVariation::Initial => {
-                Some(Self::Original)
-            }
-            EventVariation::Stripped => Some(Self::PossiblyRedacted),
-            EventVariation::Original
-            | EventVariation::OriginalSync
-            | EventVariation::Redacted
-            | EventVariation::RedactedSync => None,
-        }
-    }
-
-    /// Get the event content trait matching this variation.
-    fn to_event_content_trait(self) -> EventContentTraitVariation {
-        match self {
-            Self::Original => EventContentTraitVariation::Original,
-            Self::PossiblyRedacted => EventContentTraitVariation::PossiblyRedacted,
-        }
-    }
-}
-
-impl fmt::Display for EventContentVariation {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Original => Ok(()),
-            Self::PossiblyRedacted => write!(f, "PossiblyRedacted"),
-        }
     }
 }